Move 'Create OoS bed' test to future_manager folder
As mentioned in the previous commit we are moving tests from the
outOfServiceBeds.cy.ts file to their own files based around
functionality housed within a folder named after the role which is
allowed to peform the activity.
